+++ packages/js-sdk/src/api/index.ts
@@ -59,7 +59,7 @@ class ApiClient {
   ) {
     if (opts?.requireApiKey && !config.apiKey) {
       throw new AuthenticationError(
-        'API key is required, please visit the Team tab at https://grotte.parallactic.fr/dashboard to get your API key. ' +
+        'API key is required, please visit the Team tab at https://app.grotte.parallactic.fr/dashboard to get your API key. ' +
           'You can either set the environment variable `GROTTE_API_KEY` ' +
           "or you can pass it directly to the sandbox like Sandbox.create({ apiKey: 'grt_...' })"
       )
@@ -67,7 +67,7 @@ class ApiClient {
 
     if (opts?.requireAccessToken && !config.accessToken) {
       throw new AuthenticationError(
-        'Access token is required, please visit the Personal tab at https://grotte.parallactic.fr/dashboard to get your access token. ' +
+        'Access token is required, please visit the Personal tab at https://app.grotte.parallactic.fr/dashboard to get your access token. ' +
           'You can set the environment variable `GROTTE_ACCESS_TOKEN` or pass the `accessToken` in options.'
       )
     }