+++ packages/python-sdk/grotte/api/__init__.py
@@ -102,7 +102,7 @@ class ApiClient(AuthenticatedClient):
         if require_api_key:
             if config.api_key is None:
                 raise AuthenticationException(
-                    "API key is required, please visit the Team tab at https://grotte.parallactic.fr/dashboard to get your API key. "
+                    "API key is required, please visit the Team tab at https://app.grotte.parallactic.fr/dashboard to get your API key. "
                     "You can either set the environment variable `GROTTE_API_KEY` "
                     'or you can pass it directly to the method like api_key="grt_..."',
                 )
@@ -111,7 +111,7 @@ class ApiClient(AuthenticatedClient):
         if require_access_token:
             if config.access_token is None:
                 raise AuthenticationException(
-                    "Access token is required, please visit the Personal tab at https://grotte.parallactic.fr/dashboard to get your access token. "
+                    "Access token is required, please visit the Personal tab at https://app.grotte.parallactic.fr/dashboard to get your access token. "
                     "You can set the environment variable `GROTTE_ACCESS_TOKEN` or pass the `access_token` in options.",
                 )
             token = config.access_token